Refactor clipboard interface to own file
